refactor: refactor knockout component out.
