+++ src/pathwayvis/components/knockout/knockout.component.html
