--- src/pathwayvis/components/sidebar/sections/views/knockout.section.scss
@@ -1,1 +1,0 @@
-
